Rename to generateSuggestion for clear intention.
