+++ src/features/suggestions/index.ts
@@ -1,2 +1,2 @@
 export type { Suggestion } from "./Suggestion";
-export { generateSuggestions } from "./search";
+export { generateSuggestions } from "./search/generateSuggestion";
