+++ src/features/suggestions/search/generateSuggestion.ts
